+++ R/platjam-coverhm.R
@@ -8,7 +8,9 @@
 
 #' Make coverage heatmaps
 #'
-#' Make coverage heatmaps using a simple set of config options
+#' (IN DEV) Make coverage heatmaps using a simple set of config options
+#'
+#' This function is still in development and is not yet active.
 #'
 #' This function is intended as a wrapper function which calls
 #' `nmatlist2heatmaps()` in a more organized way.