+++ NEWS.md
@@ -9,6 +9,11 @@ the two directions are visibly distinct even under three
 different color blindness simulations from
 `dichromat::dichromat()`.
 
+## changes to existing functions
+
+* `nmatlist2heatmaps()` changes that forced heatmap legend direction
+to be `"vertical" were reverted, since Dr. Gu updated
+`ComplexHeatmap` to fix the issue with `grid` on R version 3.6.*.
 
 # platjam 0.0.33.900